packs/generate/requirements.txt
distro==1.9.0
mergedeep==1.3.4
PyYAML==6.0.2
ruamel.yaml==0.18.6
ruyaml==0.91.0
setuptools==75.6.0
Unidecode==1.3.8
